Refactor review screen directory structure
