+++ apps/web/src/screens/review/components/useReviewCardEditor.ts
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import type { TranslationKey } from "../../i18n";
-import { captureAppOperationError } from "../../observability/appOperationObservation";
-import { toCardFormState, type CardFormState } from "../cards/CardForm";
-import type { Card } from "../../types";
+import type { TranslationKey } from "../../../i18n";
+import { captureAppOperationError } from "../../../observability/appOperationObservation";
+import { toCardFormState, type CardFormState } from "../../cards/CardForm";
+import type { Card } from "../../../types";
 
 type UseReviewCardEditorParams = Readonly<{
   deleteCardItem: (cardId: string) => Promise<Card>;
