+++ apps/web/src/screens/review/data/useReviewScreenData.submit.test.tsx
@@ -2,15 +2,15 @@
 import { act } from "react";
 import ReactDOM from "react-dom/client";
 import { describe, expect, it } from "vitest";
-import type { Card, ReviewQueueSnapshot } from "../../types";
-import { I18nProvider } from "../../i18n";
+import type { Card, ReviewQueueSnapshot } from "../../../types";
+import { I18nProvider } from "../../../i18n";
 import {
   createCard,
   createDeferredPromise,
   loadReviewQueueSnapshotMock,
   ReviewScreenDataHarness,
   setupReviewScreenTest,
-} from "./ReviewScreenTestSupport";
+} from "../testSupport/ReviewScreenTestSupport";
 import {
   type ReviewSubmissionOutcome,
   type UseReviewScreenDataResult,
