+++ apps/web/src/screens/review/hardReminder/ReviewHardReminderDialog.tsx
@@ -1,5 +1,5 @@
 import type { ReactElement } from "react";
-import { useI18n } from "../../i18n";
+import { useI18n } from "../../../i18n";
 
 export type ReviewHardReminderDialogProps = Readonly<{
   isOpen: boolean;
